requirements.txt
certifi==2020.4.5.1
chardet==3.0.4
dj-database-url==0.5.0
Django==2.2.4
django-cors-headers==3.3.0
django-timezone-utils==0.13
djangorestframework==3.11.0
djangorestframework-jwt==1.11.0
gunicorn==20.0.4
idna==2.9
psycopg2-binary==2.8.3
PyJWT==1.7.1
pytz==2019.2
requests==2.23.0
sqlparse==0.3.0
urllib3==1.25.9